+++ Adafruit_IO/errors.py
@@ -18,6 +18,17 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+
+import json
+
+# MQTT RC Error Types
+MQTT_ERRORS   = [ 'Connection successful',
+                  'Incorrect protocol version',
+                  'Invalid Client ID',
+                  'Server unavailable ',
+                  'Bad username or password',
+                  'Not authorized' ]
+
 class AdafruitIOError(Exception):
     """Base class for all Adafruit IO request failures."""
     pass
@@ -26,8 +37,16 @@ class AdafruitIOError(Exception):
 class RequestError(Exception):
     """General error for a failed Adafruit IO request."""
     def __init__(self, response):
-        super(RequestError, self).__init__("Adafruit IO request failed: {0} {1}".format(
-            response.status_code, response.reason))
+        error_message = self._parse_error(response)
+        super(RequestError, self).__init__("Adafruit IO request failed: {0} {1} - {2}".format(
+            response.status_code, response.reason, error_message))
+
+    def _parse_error(self, response):
+        try:
+            content = json.loads(response.content)
+            return ' - '.join(content['error'])
+        except ValueError:
+            return ""
 
 
 class ThrottlingError(AdafruitIOError):
@@ -38,3 +57,12 @@ class ThrottlingError(AdafruitIOError):
         super(ThrottlingError, self).__init__("Exceeded the limit of Adafruit IO "  \
             "requests in a short period of time. Please reduce the rate of requests " \
             "and try again later.")
+
+
+class MQTTError(Exception):
+    """Handles connection attempt failed errors.
+    """
+    def __init__(self, response):
+        error = MQTT_ERRORS[response]
+        super(MQTTError, self).__init__(error)
+    pass