+++ Adafruit_IO/mqtt_client.py
@@ -21,7 +21,8 @@
 import logging
 
 import paho.mqtt.client as mqtt
-
+import sys
+from .errors import MQTTError, RequestError
 
 # How long to wait before sending a keep alive (paho-mqtt configuration).
 KEEP_ALIVE_SEC = 60  # One minute
@@ -34,23 +35,31 @@ class MQTTClient(object):
     using the MQTT protocol.
     """
 
-    def __init__(self, username, key, service_host='io.adafruit.com', service_port=1883):
+    def __init__(self, username, key, service_host='io.adafruit.com', secure=True):
         """Create instance of MQTT client.
 
-        Required parameters:
-        - username: The Adafruit.IO username for your account (found on the
-                    accounts site https://accounts.adafruit.com/).
-        - key: The Adafruit.IO access key for your account.
+            :param username: Adafruit.IO Username for your account.
+            :param key: Adafruit IO access key (AIO Key) for your account.
+            :param secure: (optional, boolean) Switches secure/insecure connections
         """
         self._username = username
         self._service_host = service_host
-        self._service_port = service_port
+        if secure:
+            self._service_port = 8883
+        elif not secure:
+            self._service_port = 1883
         # Initialize event callbacks to be None so they don't fire.
         self.on_connect    = None
         self.on_disconnect = None
         self.on_message    = None
         # Initialize MQTT client.
         self._client = mqtt.Client()
+        if secure:
+            self._client.tls_set_context()
+            self._secure = True
+        elif not secure:
+            print('**THIS CONNECTION IS INSECURE** SSL/TLS not supported for this platform')
+            self._secure = False
         self._client.username_pw_set(username, key)
         self._client.on_connect    = self._mqtt_connect
         self._client.on_disconnect = self._mqtt_disconnect
@@ -62,11 +71,12 @@ class MQTTClient(object):
         # Check if the result code is success (0) or some error (non-zero) and
         # raise an exception if failed.
         if rc == 0:
+            #raise RequestError(rc)
             self._connected = True
+            print('Connected to Adafruit IO!')
         else:
-            # TODO: Make explicit exception classes for these failures:
-            # 0: Connection successful 1: Connection refused - incorrect protocol version 2: Connection refused - invalid client identifier 3: Connection refused - server unavailable 4: Connection refused - bad username or password 5: Connection refused - not authorised 6-255: Currently unused.
-            raise RuntimeError('Error connecting to Adafruit IO with rc: {0}'.format(rc))
+            # handle RC errors within `errors.py`'s MQTTError class
+            raise MQTTError(rc)
         # Call the on_connect callback if available.
         if self.on_connect is not None:
             self.on_connect(self)
@@ -78,7 +88,8 @@ class MQTTClient(object):
         # log the RC as an error.  Continue on to call any disconnect handler
         # so clients can potentially recover gracefully.
         if rc != 0:
-            logger.debug('Unexpected disconnect with rc: {0}'.format(rc))
+            raise MQTTError(rc)
+        print('Disconnected from Adafruit IO!')
         # Call the on_disconnect callback if available.
         if self.on_disconnect is not None:
             self.on_disconnect(self)
@@ -91,8 +102,10 @@ class MQTTClient(object):
         if self.on_message is not None and self._username == parsed_topic[0]:
             feed = parsed_topic[2]
             payload = '' if msg.payload is None else msg.payload.decode('utf-8')
-            retain = msg.retain
-            self.on_message(self, feed, payload, retain)
+        elif self.on_message is not None and parsed_topic[0] == 'time':
+            feed = parsed_topic[0]
+            payload = msg.payload.decode('utf-8')
+        self.on_message(self, feed, payload)
 
     def connect(self, **kwargs):
         """Connect to the Adafruit.IO service.  Must be called before any loop
@@ -154,6 +167,22 @@ class MQTTClient(object):
         the on_message function will be called with the feed_id and new value.
         """
         self._client.subscribe('{0}/feeds/{1}'.format(self._username, feed_id))
+
+    def subscribe_time(self, time):
+        """Subscribe to changes on the Adafruit IO time feeds. When the feed is
+        updated, the on_message function will be called and publish a new value:
+        time =
+            millis: milliseconds
+            seconds: seconds
+            iso: ISO-8601 (https://en.wikipedia.org/wiki/ISO_8601)
+        """
+        if time == 'millis' or time == 'seconds':
+            self._client.subscribe('time/{0}'.format(time))
+        elif time == 'iso':
+            self._client.subscribe('time/ISO-8601')
+        else:
+            print('ERROR: Invalid time type specified')
+            return
 
     def publish(self, feed_id, value):
         """Publish a value to a specified feed.