+++ tests/test_mqtt_client.py
@@ -19,6 +19,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 import time
+import unittest
 
 from Adafruit_IO import MQTTClient
 
@@ -31,11 +32,11 @@ TIMEOUT_SEC = 5  # Max amount of time (in seconds) to wait for asyncronous event
 
 class TestMQTTClient(base.IOTestCase):
 
-    def wait_until_connected(self, client, connect_value=True, 
+    def wait_until_connected(self, client, connect_value=True,
         timeout_sec=TIMEOUT_SEC):
-        # Pump the specified client message loop and wait until it's connected, 
-        # or the specified timeout has ellapsed.  Can specify an explicit 
-        # connection state to wait for by setting connect_value (defaults to 
+        # Pump the specified client message loop and wait until it's connected,
+        # or the specified timeout has ellapsed.  Can specify an explicit
+        # connection state to wait for by setting connect_value (defaults to
         # waiting until connected, i.e. True).
         start = time.time()
         while client.is_connected() != connect_value and \
@@ -49,8 +50,10 @@ class TestMQTTClient(base.IOTestCase):
         # Verify not connected by default.
         self.assertFalse(client.is_connected())
 
-    def test_connect(self):
-        # Create MQTT test client.
+    def test_secure_connect(self):
+        """Test a secure (port 8883, TLS enabled) AIO connection
+        """
+        # Create MQTT-Secure test client.
         client = MQTTClient(self.get_test_username(), self.get_test_key())
         # Verify on_connect handler is called and expected client is provided.
         def on_connect(mqtt_client):
@@ -61,6 +64,25 @@ class TestMQTTClient(base.IOTestCase):
         self.wait_until_connected(client)
         # Verify connected.
         self.assertTrue(client.is_connected())
+        self.assertTrue(client._secure)
+
+    def test_insecure_connect(self):
+        """Test an insecure (port 1883, TLS disabled) AIO connection
+        """
+        # Create MQTT-Insecure (non-SSL) test client.
+        client = MQTTClient(self.get_test_username(), self.get_test_key(), secure=False)
+        # Verify on_connect handler is called and expected client is provided.
+        def on_connect(mqtt_client):
+            self.assertEqual(mqtt_client, client)
+        client.on_connect = on_connect
+        # Connect and wait until on_connect event is fired.
+        client.connect()
+        self.wait_until_connected(client)
+        # Verify connected.
+        self.assertTrue(client.is_connected())
+        # Verify insecure connection established
+        self.assertFalse(client._secure)
+
 
     def test_disconnect(self):
         # Create MQTT test client.
@@ -91,13 +113,13 @@ class TestMQTTClient(base.IOTestCase):
         client.connect()
         self.wait_until_connected(client)
         # Subscribe to changes on a feed.
-        client.subscribe('TestFeed')
+        client.subscribe('testfeed')
         # Publish a message on the feed.
-        client.publish('TestFeed', 42)
+        client.publish('testfeed', 42)
         # Wait for message to be received or timeout.
         start = time.time()
         while len(messages) == 0 and (time.time() - start) < TIMEOUT_SEC:
             client.loop()
             time.sleep(0)
         # Verify one update message with payload is received.
-        self.assertListEqual(messages, [('TestFeed', '42')])
+        self.assertListEqual(messages, [('testfeed', '42')])